ccvt_checks_works_approve_disapprove

+++ database/migrations/2020_12_11_074953_create_payments_table.php
@@ -17,7 +17,7 @@ class CreatePaymentsTable extends Migration
             $table->id();
             $table->integer('license_id');
             $table->integer('sales_person_id');
-            $table->decimal('commission',2);
+            $table->bigInteger('commission');
             $table->boolean('is_approved');
             $table->timestamps();
         });
